EIN-4504 fiksa PR kommentarar

+++ src/main/java/no/einnsyn/apiv3/entities/tilbakemelding/TilbakemeldingRepository.java
@@ -3,5 +3,5 @@ package no.einnsyn.apiv3.entities.tilbakemelding;
 import no.einnsyn.apiv3.entities.EinnsynRepository;
 import no.einnsyn.apiv3.entities.tilbakemelding.models.Tilbakemelding;
 
-public interface TilbakemeldingRepository extends EinnsynRepository<Tilbakemelding, Long> {
+public interface TilbakemeldingRepository extends EinnsynRepository<Tilbakemelding, Integer> {
 }
